pyproject.toml
[project]
name = "tarot-scan"
version = "0.1.0"
description = "Flatbed scanner + OpenCV + VLM tarot card classification toolkit"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
tarot-scan = "tarot_scan.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tarot_scan"]

[tool.ruff]
line-length = 100
target-version = "py311"
